### 2018-03-28 #655 @whedon
[REVIEW]: Fast, Consistent Tokenization of Natural Language Text
**Submitting author:** @lmullen (<a href="http://orcid.org/0000-0001-5103-6917">Lincoln Mullen</a>)
**Repository:** <a href="https://github.com/ropensci/tokenizers" target ="_blank">https://github.com/ropensci/tokenizers</a>
**Version:** v0.2.0
