### 2018-03-28 #655 @arfon
Accepting because of https://github.com/ropensci/onboarding/issues/33
