### 2018-03-28 #655 @lmullen
@arfon Wow. That was fast.

I noticed just a few minor formatting things with the BibTex. If I just update `paper.md` and `paper.bib`, can I regenerate the published version?
