### 2018-03-28 #655 @arfon
> I noticed just a few minor formatting things with the BibTex. If I just update paper.md and paper.bib, can I regenerate the published version?

Sure thing. Please just update the files in `master` on https://github.com/ropensci/tokenizers and I can update.
