### 2018-03-29 #655 @lmullen
@arfon I've corrected some typos and the BibTeX for this paper as well. Could you please regenerate the final version? I didn't realize we could generate a PDF proof with the bot, or I would have done this earlier.
