### 2018-04-05 #655 @lmullen
@arfon Sorry to be a pest. But I noticed that the published PDF did not update after you regenerated it. I.e. the proof a few lines above is corrected but the published version isnt. Would you mind updating it if possible?
