### 2018-04-06 #655 @arfon
> @arfon Sorry to be a pest. But I noticed that the published PDF did not update after you regenerated it. I.e. the proof a few lines above is corrected but the published version isnt. Would you mind updating it if possible?

@lmullen - could you tell me what difference I should be looking for and can you confirm that this paper is _not_ correct: https://github.com/openjournals/joss-papers/blob/5716811674dbd048372cd310e8598708e82a235b/joss.00655/10.21105.joss.00655.pdf ?
