### 2018-04-06 #655 @lmullen
@arfon Yes, the [link above](https://github.com/openjournals/joss-papers/blob/5716811674dbd048372cd310e8598708e82a235b/joss.00655/10.21105.joss.00655.pdf) is not correct. If you compare that to [this proof](https://github.com/openjournals/joss-papers/blob/joss.00655/joss.00655/10.21105.joss.00655.pdf) you can check that the link to tokenizers in the second paragraph should *not* be bold and that in the bibliography my name should only appear once as `Mullen, Lincoln A. ` rather than two different entries, among other minor corrections.
